requirements.txt
streamlit
pandas
python-dotenv
openai
bcrypt
psycopg2-binary
